perf: sort table comluns
