+++ package.json
@@ -38,6 +38,7 @@
     "deepmerge": "^4.2.2",
     "dompurify": "^3.1.6",
     "echarts": "4.7.0",
+    "el-table-draggable": "^1.4.12",
     "element-ui": "2.15.14",
     "eslint-plugin-html": "^6.0.0",
     "highlight.js": "^11.9.0",
@@ -67,6 +68,7 @@
     "npm": "^7.8.0",
     "nprogress": "0.2.0",
     "path-to-regexp": "3.3.0",
+    "sortablejs": "^1.15.6",
     "v-sanitize": "^0.0.13",
     "vue": "2.6.10",
     "vue-codemirror": "4.0.6",
